+++ lib/src/model/annotation.dart
@@ -7,7 +7,7 @@ import 'package:dartdoc/src/element_type.dart';
 import 'package:dartdoc/src/model/feature.dart';
 import 'package:dartdoc/src/model/getter_setter_combo.dart';
 import 'package:dartdoc/src/model/library.dart';
-import 'package:dartdoc/src/model/model_element_builder.dart';
+import 'package:dartdoc/src/model/model_object_builder.dart';
 import 'package:dartdoc/src/model/package_graph.dart';
 
 /// Represents a Dart annotation, attached to an element in the source code with
