+++ test/comment_referable/comment_referable_test.dart
@@ -6,7 +6,7 @@ library dartdoc.comment_reference_test;
 
 import 'package:analyzer/dart/element/element.dart';
 import 'package:dartdoc/src/model/comment_referable.dart';
-import 'package:dartdoc/src/model/model_element_builder.dart';
+import 'package:dartdoc/src/model/model_object_builder.dart';
 import 'package:dartdoc/src/model/nameable.dart';
 import 'package:test/test.dart';
